# backend/tests/expected/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CalculateCostRequest": {
        "properties": {
          "absorber_catalog_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Absorber Catalog Id"
          },
          "absorber_markup": {
            "default": 100,
            "title": "Absorber Markup",
            "type": "integer"
          },
          "amp_catalog_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amp Catalog Id"
          },
          "amp_markup": {
            "default": 100,
            "title": "Amp Markup",
            "type": "integer"
          },
          "catalog_id": {
            "title": "Catalog Id",
            "type": "string"
          },
          "damage_enhancers": {
            "default": 0,
            "title": "Damage Enhancers",
            "type": "integer"
          },
          "scope_catalog_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Catalog Id"
          },
          "scope_markup": {
            "default": 100,
            "title": "Scope Markup",
            "type": "integer"
          },
          "type": {
            "default": "weapon",
            "enum": [
              "weapon",
              "healing"
            ],
            "title": "Type",
            "type": "string"
          },
          "weapon_markup": {
            "default": 100,
            "title": "Weapon Markup",
            "type": "integer"
          }
        },
        "required": [
          "catalog_id"
        ],
        "title": "CalculateCostRequest",
        "type": "object"
      },
...
      "CostResult": {
        "additionalProperties": true,
        "description": "A standalone cost calculation (breakdown plus total).",
        "properties": {
          "costBreakdown": {
            "items": {
              "$ref": "#/components/schemas/CostBreakdownLine"
            },
            "title": "Costbreakdown",
            "type": "array"
          },
          "totalCostPerUse": {
            "title": "Totalcostperuse",
            "type": "number"
          }
        },
        "required": [
          "costBreakdown",
          "totalCostPerUse"
        ],
        "title": "CostResult",
        "type": "object"
      },
      "DeletedStatus": {
        "additionalProperties": true,
        "description": "A ``{\"status\": \"deleted\"}`` soft-delete acknowledgement.",
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DeletedStatus",
        "type": "object"
      },
      "Equipment": {
        "additionalProperties": true,
        "description": "A library entry with computed per-use cost (the list shape).",
        "properties": {
          "amplifierName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amplifiername"
          },
          "costPerUse": {
            "title": "Costperuse",
            "type": "number"
          },
          "damageMax": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Damagemax"
          },
          "damageMin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Damagemin"
          },
          "enrichmentLevel": {
            "title": "Enrichmentlevel",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "isLimited": {
            "title": "Islimited",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reloadSeconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reloadseconds"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "costPerUse",
          "isLimited",
          "enrichmentLevel"
        ],
        "title": "Equipment",
        "type": "object"
      },
...
      "EquipmentSearchResult": {
        "additionalProperties": true,
        "description": "A catalogue search hit, with ammo burn already converted to PEC.",
        "properties": {
          "ammoBurn": {
            "title": "Ammoburn",
            "type": "number"
          },
          "catalogId": {
            "title": "Catalogid",
            "type": "string"
          },
          "decay": {
            "title": "Decay",
            "type": "number"
          },
          "isLimited": {
            "title": "Islimited",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "catalogId",
          "name",
          "decay",
          "ammoBurn",
          "isLimited"
        ],
        "title": "EquipmentSearchResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EntropiaOrme API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/equipment/cost/calculate": {
      "post": {
        "description": "Calculate cost breakdown for a weapon configuration without saving.",
        "operationId": "calculate_cost_api_equipment_cost_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculateCostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Cost",
        "tags": [
          "equipment"
        ]
      }
    },
    "/api/equipment/library": {
      "get": {
        "description": "Return all items in the user's equipment library with computed costs.",
        "operationId": "get_library_api_equipment_library_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Equipment"
                  },
                  "title": "Response Get Library Api Equipment Library Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Library",
        "tags": [
          "equipment"
        ]
      },
      "post": {
        "description": "Add an item to the equipment library.",
        "operationId": "add_to_library_api_equipment_library_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWeaponRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add To Library",
        "tags": [
          "equipment"
        ]
      }
    },
    "/api/equipment/library/{item_id}": {
      "delete": {
        "description": "Remove an item from the equipment library.",
        "operationId": "remove_from_library_api_equipment_library__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove From Library",
        "tags": [
          "equipment"
        ]
      },
      "put": {
        "description": "Update an existing equipment library entry.",
        "operationId": "update_library_item_api_equipment_library__item_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWeaponRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Library Item",
        "tags": [
          "equipment"
        ]
      }
    },
    "/api/equipment/library/{item_id}/detail": {
      "get": {
        "description": "Return full detail with cost breakdown for a library item.",
        "operationId": "get_library_detail_api_equipment_library__item_id__detail_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Library Detail",
        "tags": [
          "equipment"
        ]
      }
    },
    "/api/equipment/search": {
      "get": {
        "description": "Search the bundled equipment catalogue by name and type.\n\ntype: weapon | amp | healer | scope | absorber | consumable\nReturns items with ammoBurn already converted to PEC.",
        "operationId": "search_items_api_equipment_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "default": "weapon",
              "title": "Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EquipmentSearchResult"
                  },
                  "title": "Response Search Items Api Equipment Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Items",
        "tags": [
          "equipment"
        ]
      }
    },
...
  }
}